+++ src/ru/gb/family_tree/exception/HumanExcistsException.java
@@ -1,4 +1,4 @@
-package ru.gb.family_tree;
+package ru.gb.family_tree.exception;
 
 public class HumanExcistsException extends Exception{
     public HumanExcistsException(String message) {
